requirements.txt
astroid==3.2.4
coverage==7.6.1
dill==0.3.8
exceptiongroup==1.2.2
Faker==30.1.0
iniconfig==2.0.0
isort==5.13.2
mccabe==0.7.0
packaging==24.1
platformdirs==4.3.3
pluggy==1.5.0
pylint==3.2.7
pytest==8.3.3
pytest-cov==5.0.0
pytest-pylint==0.21.0
python-dateutil==2.9.0.post0
python-dotenv
six==1.16.0
tomli==2.0.1
tomlkit==0.13.2
typing_extensions==4.12.2
